@@ -21,7 +21,13 @@ namespace AspNetCoreExample
         public static IWebHostBuilder CreateWebHostBuilder(string[] args)
         {
             var result = WebHost.CreateDefaultBuilder(args)
-                  .UseStartup(Assembly.GetEntryAssembly().FullName);
+                  .UseStartup(Assembly.GetEntryAssembly().FullName)
+                  .ConfigureLogging((hostingContext, logging) =>
+                  {
+                      logging.AddConfiguration(hostingContext.Configuration.GetSection("Logging"));
+                      logging.AddConsole();
+                      logging.AddDebug();
+                  });
             
             var port = Environment.GetEnvironmentVariable("PORT");
 
